Refactor into modular parts
