+++ ftl/integration-sample/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "integration-sample",
   "version": "1.0.0",
   "type": "module",
   "description": "",
